This is a bit of a messy change.  The goal here is to have the request object of the system available to the bottom level RPCs that are being called.
There are various reasons for this; mine being that I want to know
the IP address and other details of the caller.

Now... this is a messy wan to do it, but ... hey... it works.  It
adds an argument to the manager's "handle" routine (among others) to
carry the request object.

For the low level functions, the object arrives as a keyword
argument 'request_obj'.

Other json-rpc frameworks seem to have the signature of the low
level functions include a request object.  This is not astethically
pleasing, but it is a functional way to do it.  It does mean
that you can't RPC out lambdas easily, but that's of limited
use in the real world.

If you don't want to make this all non-thread-safe, I don't see
too many other solutions.